Added support for root CBVs in the rootsignature

+++ Runtime/VulkanRHI/VulkanPipelineLayout.h
@@ -44,7 +44,7 @@ struct FVulkanDescriptorRemappingInfo
     {
         EVulkanBindingType BindingType;
         uint8        BindingIndex;
-        uint8        OriginalBindingIndex;
+        uint16       OriginalBindingIndex;
     };
 
     FVulkanDescriptorRemappingInfo()
